image to int image

+++ free_loc/main.py
@@ -253,7 +253,7 @@ def main():
                     img = input[j].cpu().numpy()
                     img = np.transpose(img, (1,2,0))
 
-                    heatmapimage_ = int(heatmapimage_.cpu().detach().numpy()*255)
+                    heatmapimage_ = (heatmapimage_.cpu().detach().numpy()*255).astype(int)
                     ret,thresh = cv2.threshold(img,200,255,0)
                     contours,hierarchy = cv2.findContours(thresh, 1, 2)
